refactor(sheet): remove dead bottom-sheet ViewModelBase paths
